Modified architecture.
Extracted common code and created a class with it (showpins).
Created a class that inherite from showpins for each new board.
The new xxxpins.py contains only the description of the board and a call
to the display method of showpins. It makes it easier to create a new
board starting from any already created one.